+++ src/organization/Measure.ts
@@ -15,7 +15,7 @@ export class Measure extends RaainNode {
     constructor(json: {
         id: string;
         values: IPolarMeasureValue[] | ICartesianMeasureValue[] | number[];
-        date?: Date;
+        date?: Date | string;
         validity?: number;
         configurationAsJSON?: string;
         version?: string;