+++ src/quality/history/CartesianRainHistory.ts
@@ -4,7 +4,7 @@ export class CartesianRainHistory {
     public date: Date;
     public computedValue: CartesianValue;
 
-    constructor(json: {date: Date; computedValue: CartesianValue}) {
+    constructor(json: {date: Date | string; computedValue: CartesianValue}) {
         this.date = json.date ? new Date(json.date) : undefined;
         this.computedValue = json.computedValue;
     }