+++ src/radar/RadarMeasure.ts
@@ -15,7 +15,7 @@ export class RadarMeasure extends Measure {
     constructor(json: {
         id: string;
         values: IPolarMeasureValue[] | ICartesianMeasureValue[] | number[];
-        date?: Date;
+        date?: Date | string;
         validity?: number;
         configurationAsJSON?: string;
         radar?: string;