+++ src/rain/RainComputationMap.ts
@@ -53,7 +53,7 @@ export class RainComputationMap extends RainComputationAbstract {
      */
     constructor(json: {
         id: string;
-        date: Date;
+        date: Date | string;
         isReady: boolean;
         map: RainMeasure[] | string;
         links?: Link[] | RaainNode[];
@@ -62,7 +62,7 @@ export class RainComputationMap extends RainComputationAbstract {
         progressIngest?: number;
         progressComputing?: number;
         timeSpentInMs?: number;
-        isDoneDate?: Date;
+        isDoneDate?: Date | string;
         launchedBy?: string;
         rain?: string | Link | RaainNode;
         radars?: string[] | Link[] | RaainNode[];