+++ src/rain/RainMeasure.ts
@@ -12,7 +12,7 @@ export class RainMeasure extends Measure {
     constructor(json: {
         id: string;
         values: IPolarMeasureValue[] | ICartesianMeasureValue[] | number[];
-        date?: Date;
+        date?: Date | string;
         validity?: number;
         configurationAsJSON?: string;
     }) {